+++ duckbot/slash/__init__.py
@@ -1,8 +1,8 @@
 from .context import InteractionContext
-from .core import SlashCommandPatch
 from .interaction import Interaction
 from .option import Option, OptionType
 from .slash_command import slash_command
+from .slash_command_patch import SlashCommandPatch
 
 
 def setup(bot):
