+++ duckbot/slash/slash_command_patch.py
